+++ js/sr.js
@@ -433,7 +433,7 @@ $(document).ready(function() {
       if ( a != "" && a != "Outro..." ) { a = "# HD: " + a + "\n\n"; }
       else { a = ""; }
       sr_header = $("#sr-header").val();
-      $("#output-uni").val(sr_header + "\n" + s + "\n\n" + o + "\n\n" + a + "# CD:\n" + p);
+      $("#output-uni").val("# " + sr_header + " #\n" + s + "\n\n" + o + "\n\n" + a + "# CD:\n" + p);
     }
   });
 
@@ -845,7 +845,7 @@ function runP() {
   if( $("#plan-med2").is(":checked") ) { p.push( "Prescrevo " + $("#med2").val() + " para " + $("#med2-reason").val() ); }
   if( $("#plan-med3").is(":checked") ) { p.push( "Prescrevo " + $("#med3").val() + " para " + $("#med3-reason").val() ); }
   // Guidance
-  if( $("#plan-warnsigns").is(":checked") ) { p.push( "Oriento sinais de alarme para retorno precoce e reavaliação em acolhimento no Sintomáticos Respiratórios" ); }
+  if( $("#plan-warnsigns").is(":checked") ) { p.push( "Oriento sinais de alarme para retorno precoce e reavaliação" ); }
   p.push( "Oriento ausência de evidências científicas validando o uso de \"tratamento precoce\" para COVID-19 (e.g. dexametasona, ivermectina, azitromicina) em pacientes ambulatoriais, assim como reforço a CONTRAINDICAÇÃO FORMAL de uso de azitromicina em casos de suspeita de infeção por SARS-CoV-2 sem sinais de infecção bacteriana (Choosing Wisely Brasil)" );
   if( $("#plan-isolation").is(":checked") ) { p.push( "Oriento sobre medidas de isolamento do caso-índice tanto dentro como fora de casa (uso de máscara durante todo o período, dormir e se manter em cômodo separado, manter casa arejada e ventilada, não compartilhar itens pessoais, de higiene ou de alimentação, etc.), higiene das mãos e medidas de etiqueta respiratória" ); }
   if( $("#plan-quarantine").is(":checked") ) { p.push( "Oriento medidas que as demais pessoas da casa devem tomar para evitar entrar em contato com o caso índice, assim como reforço higiene de mãos e etiqueta respiratória" ); }